Remove the loop's throw path, fix nine verified defects, sweep every animation
The loop had exactly one way to crash that nothing guarded. Both history
buffers were std::deque, so AudioHistoryTracker and MoodHistory took and
returned a 480-byte and a 456-byte node about fifteen times a second, and a
failed push_back threw std::bad_alloc out of loop() into std::terminate. The
existing heap gate did not cover it, because free heap is a total while the
allocation needs one contiguous block. Both are now SnapshotRing, a
fixed-capacity ring in src/audio/SnapshotRing.h. addSnapshot allocates zero
times across a whole run and the 456 and 480 byte rows are gone from the
harness's size histogram.

The three try/catch blocks in LayerManager guarded only fadeToBlackBy and
memcpy, none of which throw, while the allocations in the same function sat
outside them. They are removed rather than left implying protection that was
never there, and loop() now has a backstop catch.

Nine further defects are fixed. The scene picker could return the scene
already running, which reset the clock and made the transition invisible.
SceneDefinition::activeScene was dead. A C++20 designated initializer in
AudioHistoryTracker compiled only as a GCC extension under gnu++11. A
function-local static scratch buffer in LayerManager was shared across strips.
neonFlow underflowed its brightness for roughly 5 of every 16 pixels.
AlienPulse passed a negative float to a uint8_t. Six float phase accumulators
grew without bound. All eight catalog animations indexed ahead of a
zero-length guard. And a heap monitor with no call sites is gone.

The host harness grew from 14 checks to 101. It sweeps all eight catalog
animations at n=100 and n=8, all eight reachable layer factories, a
device-scale audio timeline with a populated spectrum and waveform, and soaks
the four reachable accumulator sites for 20000 frames each, asserting each
stays inside one period. It measures peak history bytes rather than
reconciling them: AudioHistory peaks at 1500 elements and 60000 bytes,
MoodHistory at 150 and 11400.

Neither environment needs a board attached and neither sets upload_port, so
default_envs makes a bare `pio run` build the firmware, and platformio.ini
names both commands.

Two defects surfaced while fixing the others and are recorded in BACKLOG.md
rather than fixed: WormholeVortexLayer and CentroidColorFlowLayer are never
instantiated anywhere in src/, and the mood classifier's thresholds are
written against a 0..1 energy range while the device field is a raw FFT
magnitude sum in the thousands. Nothing has run on hardware yet.

+++ src/audio/AudioFeatures.h
@@ -28,7 +28,7 @@ struct AudioFeatures {
 
     int16_t* waveform = nullptr;    // Pointer to time-domain samples
     size_t waveformSize = 0;        // Size of waveform buffer
-    double spectrum[NUM_SAMPLES / 2] = {};  // FFT magnitudes
+    float spectrum[NUM_SAMPLES / 2] = {};   // FFT magnitudes
 
     float centroid = 0.0f;          //  ??
     float frequency = 0.0f;         //  frequency
